+++ test/reagenttest/testwrap.cljs
@@ -4,7 +4,7 @@
             [reagent.impl.template :as tmpl]
             [reagenttest.utils :as u :refer [with-mounted-component]]))
 
-(def functional-compiler (r/create-compiler {:functional-components? true}))
+(def functional-compiler (r/create-compiler {:function-components true}))
 
 (def test-compilers
   [tmpl/default-compiler*